pyproject.toml
[project]
name = "jaffle-shop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dagster-dbt>=0.26.10",
    "dagster-webserver>=1.10.10",
    "dbt-duckdb>=1.9.2",
    "duckdb>=1.2.2",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "pyarrow>=19.0.1",
]
